requirements.txt
google-ads
google-api-python-client
streamlit
openai
langchain==0.0.272
flask
flask-socketio
flask-cors
google-search-results
chromadb
sentence_transformers
tiktoken
GitPython
llama-cpp-python